Move region logic into ds, set timeout for fetching meta-data to 5 seconds.
